Refactoring trace/*Config => trace/config.go
