+++ pkg/util/trace/report_log_test.go
@@ -41,7 +41,7 @@ func TestReportLog(t *testing.T) {
 			name:         "close",
 			enableTracer: false,
 			args: args{
-				ctx:       ContextWithSpanContext(context.Background(), SpanContextWithIDs(0, 0)),
+				ctx:       ContextWithSpanContext(context.Background(), SpanContextWithIDs(nilTraceID, nilSpanID)),
 				level:     zapcore.InfoLevel,
 				depth:     3,
 				formatter: "info message",
@@ -52,7 +52,7 @@ func TestReportLog(t *testing.T) {
 			name:         "collect",
 			enableTracer: true,
 			args: args{
-				ctx:       ContextWithSpanContext(context.Background(), SpanContextWithIDs(0, 0)),
+				ctx:       ContextWithSpanContext(context.Background(), SpanContextWithIDs(nilTraceID, nilSpanID)),
 				level:     zapcore.InfoLevel,
 				depth:     3,
 				formatter: "info message",
